refactor: modularize lowrank gls core
